set balance
